requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.34,<1.0
sqlmodel>=0.0.18,<0.1.0
python-jose>=3.4
passlib[bcrypt]>=1.7,<2.0
bcrypt==3.2.0 
jinja2>=3.1,<4.0
python-multipart>=0.0.6,<1.0
